+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0).
 
 #### Added
 - `admin` Improved filter JS: added highlighting of active filters
+- `admin` Filter with autocomplete for product attributes
 
 #### Changed
 - `system` Fonts Awesome updated to version 7.3.1